+++ client-sdk/ts-web/rt/src/callformat.ts
@@ -1,6 +1,5 @@
 import * as oasis from '@oasisprotocol/client';
 import * as deoxysii from '@oasisprotocol/deoxysii';
-import * as nacl from 'tweetnacl';
 
 import * as mraeDeoxysii from './mrae/deoxysii';
 import * as transaction from './transaction';
@@ -21,10 +20,14 @@ export interface EncodeConfig {
      * publicKey is an optional runtime's call data public key to use for encrypted call formats.
      */
     publicKey?: types.KeyManagerSignedPublicKey;
+    /**
+     * epoch is the epoch of the ephemeral runtime key (when publicKey is set).
+     */
+    epoch?: oasis.types.longnum;
 }
 
 export interface MetaEncryptedX25519DeoxysII {
-    sk: Uint8Array;
+    sk: CryptoKey;
     pk: Uint8Array;
 }
 
@@ -32,14 +35,13 @@ export interface MetaEncryptedX25519DeoxysII {
  * encodeCallWithNonceAndKeys encodes a call based on its configured call format.
  * It returns the encoded call and any metadata needed to successfully decode the result.
  */
-export function encodeCallWithNonceAndKeys(
+export async function encodeCallWithNonceAndKeys(
     nonce: Uint8Array,
-    sk: Uint8Array,
-    pk: Uint8Array,
+    clientKP: CryptoKeyPair,
     call: types.Call,
     format: types.CallFormat,
     config?: EncodeConfig,
-): [types.Call, unknown] {
+): Promise<[types.Call, unknown]> {
     switch (format) {
         case transaction.CALLFORMAT_PLAIN:
             return [call, undefined];
@@ -47,9 +49,11 @@ export function encodeCallWithNonceAndKeys(
             if (config?.publicKey === undefined) {
                 throw new Error('callformat: runtime call data public key not set');
             }
+            const pk = await mraeDeoxysii.publicKeyFromKeyPair(clientKP);
+            const sk = clientKP.privateKey;
             const rawCall = oasis.misc.toCBOR(call);
             const zeroBuffer = new Uint8Array(0);
-            const sealedCall = mraeDeoxysii.boxSeal(
+            const sealedCall = await mraeDeoxysii.boxSeal(
                 nonce,
                 rawCall,
                 zeroBuffer,
@@ -61,10 +65,12 @@ export function encodeCallWithNonceAndKeys(
                 nonce: nonce,
                 data: sealedCall,
             };
+            if (config.epoch) {
+                envelope.epoch = config.epoch;
+            }
             const encoded: types.Call = {
                 format: transaction.CALLFORMAT_ENCRYPTED_X25519DEOXYSII,
-                method: '',
-                body: oasis.misc.toCBOR(envelope),
+                body: envelope,
             };
             const meta: MetaEncryptedX25519DeoxysII = {
                 sk: sk,
@@ -80,32 +86,25 @@ export function encodeCallWithNonceAndKeys(
  * encodeCall randomly generates nonce and keyPair and then call encodeCallWithNonceAndKeys
  * It returns the encoded call and any metadata needed to successfully decode the result.
  */
-export function encodeCall(
+export async function encodeCall(
     call: types.Call,
     format: types.CallFormat,
     config?: EncodeConfig,
-): [types.Call, unknown] {
+): Promise<[types.Call, unknown]> {
     const nonce = new Uint8Array(deoxysii.NonceSize);
     crypto.getRandomValues(nonce);
-    const keyPair = nacl.box.keyPair();
-    return encodeCallWithNonceAndKeys(
-        nonce,
-        keyPair.secretKey,
-        keyPair.publicKey,
-        call,
-        format,
-        config,
-    );
+    const clientKP = await mraeDeoxysii.generateKeyPair(true);
+    return await encodeCallWithNonceAndKeys(nonce, clientKP, call, format, config);
 }
 
 /**
  * decodeResult performs result decoding based on the specified call format metadata.
  */
-export function decodeResult(
+export async function decodeResult(
     result: types.CallResult,
     format: types.CallFormat,
-    meta?: MetaEncryptedX25519DeoxysII,
-): types.CallResult {
+    meta?: unknown,
+): Promise<types.CallResult> {
     switch (format) {
         case transaction.CALLFORMAT_PLAIN:
             // In case of plain-text data format, we simply pass on the result unchanged.
@@ -113,16 +112,15 @@ export function decodeResult(
         case transaction.CALLFORMAT_ENCRYPTED_X25519DEOXYSII:
             if (result.unknown) {
                 if (meta) {
-                    const envelop = oasis.misc.fromCBOR(
-                        result.unknown,
-                    ) as types.ResultEnvelopeX25519DeoxysII;
+                    const metaEncryptedX25519DeoxysII = meta as MetaEncryptedX25519DeoxysII;
+                    const envelope = result.unknown as types.ResultEnvelopeX25519DeoxysII;
                     const zeroBuffer = new Uint8Array(0);
-                    const pt = mraeDeoxysii.boxOpen(
-                        envelop?.nonce,
-                        envelop?.data,
+                    const pt = await mraeDeoxysii.boxOpen(
+                        envelope?.nonce,
+                        envelope?.data,
                         zeroBuffer,
-                        meta.pk,
-                        meta.sk,
+                        metaEncryptedX25519DeoxysII.pk,
+                        metaEncryptedX25519DeoxysII.sk,
                     );
                     return oasis.misc.fromCBOR(pt) as types.CallResult;
                 } else {